# backend/internal/dashboard/data/openapi.json
{
  "components": {
    "schemas": {
      "AuthStatusResponse": {
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "has_password": {
            "type": "boolean"
          },
          "is_default_admin_token": {
            "type": "boolean"
          },
          "require_login": {
            "type": "boolean"
          }
        },
        "required": [
          "authenticated",
          "has_password",
          "is_default_admin_token",
          "require_login"
        ],
        "type": "object"
      },
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "type": "object"
      },
      "ChangePasswordResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "message",
          "ok"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "fb_admin",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Dashboard admin surface. Generated by backend/cmd/openapi-emit from admin_manifest.json + dashboard.AdminAPIPaths — do not edit by hand.",
    "title": "freebucks-proxy admin API",
    "version": "dev"
  },
  "openapi": "3.0.3",
  "paths": {
    "/admin": {
      "get": {
        "operationId": "spaRoot",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SPA shell (HTML)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "SPA shell (HTML)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/": {
      "get": {
        "operationId": "spaRootSlash",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SPA shell (HTML)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "SPA shell (HTML)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/auth/status": {
      "get": {
        "operationId": "getAuthStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusResponse"
                }
              }
            },
            "description": "Dashboard auth state (login mode, factory-default check)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Dashboard auth state (login mode, factory-default check)",
        "tags": [
          "auth"
        ]
      }
    },
    "/admin/api/change-password": {
      "post": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            },
            "description": "Rotate the admin dashboard password"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Rotate the admin dashboard password",
        "tags": [
          "auth"
        ]
      }
    },
    "/admin/api/config": {
      "get": {
        "operationId": "getConfig",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/configData"
                }
              }
            },
            "description": "Effective config (sensitive: raw .env read)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Effective config (sensitive: raw .env read)",
        "tags": [
          "settings"
        ]
      }
    },
    "/admin/api/config/meta": {
      "get": {
        "operationId": "getConfigMeta",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getConfigMetaResponse"
                }
              }
            },
            "description": "Configuration catalog (the settings form schema)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Configuration catalog (the settings form schema)",
        "tags": [
          "settings"
        ]
      }
    },
    "/admin/api/events": {
      "get": {
        "operationId": "events",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent event stream (dashboard live updates)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Server-sent event stream (dashboard live updates)",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs": {
      "get": {
        "operationId": "getLogs",
        "parameters": [
          {
            "description": "Log level filter",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Message substring filter",
            "in": "query",
            "name": "msg",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "View window override (Go duration, clamped to 1m..168h)",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsData"
                }
              }
            },
            "description": "Recent log ring view model"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Recent log ring view model",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs/export": {
      "get": {
        "operationId": "exportLogs",
        "parameters": [
          {
            "description": "Unix-millis lower bound (default 0, clamped to 168h)",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis upper bound (default now)",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsExportDoc"
                }
              }
            },
            "description": "Streamed versioned export of request_records plus log_entries"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Streamed versioned export of request_records plus log_entries",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs/history": {
      "get": {
        "operationId": "getLogsHistory",
        "parameters": [
          {
            "description": "Log level filter",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Message substring filter",
            "in": "query",
            "name": "msg",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Request id filter",
            "in": "query",
            "name": "req_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis lower bound",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unix-millis upper bound",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Max records (default 500)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsHistoryData"
                }
              }
            },
            "description": "Persisted log records"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Persisted log records",
        "tags": [
          "views"
        ]
      }
    },
    "/admin/api/logs/import": {
      "post": {
        "operationId": "importLogs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logsImportReport"
                }
              }
            },
            "description": "Gap-fill restore of an export document (INSERT OR IGNORE, 64MB cap)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Gap-fill restore of an export document (INSERT OR IGNORE, 64MB cap)",
        "tags": [
          "views"
        ]
      }
    },
...
    "/admin/api/models": {
      "get": {
        "operationId": "getModels",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/modelsData"
                }
              }
            },
            "description": "Full model catalog view model (tiers, withdrawal and live offer state)"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Full model catalog view model (tiers, withdrawal and live offer state)",
        "tags": [
          "views"
        ]
      }
    },
...
  }
}