# backend/internal/server/server.go
// Package server exposes the OpenAI-compatible HTTP surface of the
// freebuff-proxy bridge: POST /v1/chat/completions (stream + non-stream),
// GET /v1/models, and GET /healthz. Stdlib only.
//
// Responsibilities (PRD §6 error matrix):
//   - optional client auth (Bearer / x-api-key exact match, constant-time)
//   - request sanitization via internal/convert before the upstream call
//   - retry-once recovery for session-invalid / run-invalid chat errors
//   - 30-min token cooldown on upstream auth rejection
//   - error mapping to the OpenAI error shape, 503 + Retry-After for the
//     waiting room, 502 when every token is exhausted
//   - SSE relay (sanitized chunks + [DONE]) and non-streaming accumulation
//   - client-disconnect propagation to the upstream (request context)
package server

import (
	"context"
	"fmt"
	"log/slog"
	"math"
	"net/http"
	"strings"
	"sync"
	"sync/atomic"
	"time"

	"freebucks-proxy/backend/internal/config"
	"freebucks-proxy/backend/internal/convert"
	"freebucks-proxy/backend/internal/dashboard"
	"freebucks-proxy/backend/internal/logring"
	"freebucks-proxy/backend/internal/pool"
	"freebucks-proxy/backend/internal/ratelimit"
	"freebucks-proxy/backend/internal/reasoningcache"
	"freebucks-proxy/backend/internal/registry"
	"freebucks-proxy/backend/internal/store"
	"freebucks-proxy/backend/internal/tokendb"
	"freebucks-proxy/backend/internal/tokenestimate"
	"freebucks-proxy/backend/internal/updatecheck"
	"freebucks-proxy/backend/internal/upstream"
)

const (
	// maxRequestBody caps the inbound chat-completions body (32MB).
	maxRequestBody = 32 << 20
	// maxStreamLine caps one upstream SSE line the scanner will buffer.
	maxStreamLine = 16 << 20
)

// Server is the HTTP handler holder: routes are built by Handler(). cfg is an
// atomic pointer because /admin/reload swaps it while requests are in flight;
// every read site must Load() it once per request and use the local.
type Server struct {
	cfg     atomic.Pointer[config.Config]
	pool    *pool.Pool
	reg     *registry.Registry
	logger  *slog.Logger
	started time.Time

	// logs is the optional dashboard log viewer ring (nil = disabled); its
	// Counts feed freebucks_proxy_log_events_total on /metrics.
	logs *logring.Handler

	// dash is the embedded admin UI (Svelte SPA + vendored assets).
	dash *dashboard.Dashboard
	// hist is the dashboard history store (nil = live-only views). Threaded
	// into the dashboard at construction; closed via Close on shutdown.
	hist *store.Store
	// adminAuth guards the dashboard: a stateless HMAC-signed session cookie
	// issued against ADMIN_TOKEN, plus a per-IP login rate limiter.
	adminAuth *adminAuth
	// configPath is the -config JSON path ("" when none); reloads re-apply it
	// so JSON overrides survive dashboard saves and /admin/reload.
	configPath string

	// tokenDB is the persistent SQLite-backed token store.  When non-nil,
	// add/remove mutations go through the database instead of rewriting
	// .env, and the initial token list is seeded from the database on
	// startup (with a one-time migration from AUTH_TOKENS env).
	tokenDB *tokendb.DB

	// version is the running release tag (""/dev for dev builds); the
	// dashboard badge compares it against the latest GitHub release (#50b).
	// updates is the cached latest-release checker (nil = no badge).
	// commit is the running build's commit hash (ARSYDONI UPDATE SOURCE,
	// merge-guarded): the dashboard's primary update signal compares it
	// against the repo's main-branch head. "" for dev builds.
	version string
	commit  string
	updates *updatecheck.Checker

	// authClient drives the headless OAuth login wizard (issue #62): a
	// token-less upstream client whose transport/stealth wiring matches the
	// pooled clients. nil disables the wizard endpoints with 503.
	authClient *upstream.Client
	// tokenEstimator counts tokens locally for /v1/messages/count_tokens
	// (nil only if the embedded codec failed to initialize at startup).
	tokenEstimator *tokenestimate.Estimator
	// loginFlows is the in-flight login-flow registry keyed by flow id
	// (fingerprint): start POSTs /api/auth/cli/code, status polls it until
	// the authToken lands (then AddToken + persist).
	loginFlows map[string]*loginFlow
	// reasoningCache caches reasoning content and signatures for tool calls across turns.
	reasoningCache *reasoningcache.Cache
	// rateLimiter caps client request rates per source IP (issue #137).
	rateLimiter *ratelimit.Limiter
	// rateLimitRejections tracks total client requests rejected by the
	// local rate limiter.
	rateLimitRejections atomic.Int64

	// admin owns the /admin surface (issue #250): the admin handlers are
	// methods on *adminHandlers, not *Server, so the API surface and the
	// admin surface do not share one mutable god struct.
	admin *adminHandlers

	// gates are the per-Server access-log quiescence gates (issue #252):
	// previously process-global, now owned per instance so two Servers do
	// not share one access gate.
	gates *accessGates
	// rateLimitDedupe gates identical (token, code, window) `request failed`
	// logs (D6): the first + every 50th occurrence fire; the counter always
	// increments so a silent burst stays countable, and the client response
	// is always written. Per-Server like the access gates (issue #252).
	rateLimitDedupe struct {
		mu sync.Mutex
		m  map[string]int64
	}
}

// convertOptions builds the per-request convert options from the live
// config (issue #277/#251): feature knobs resolved once (never per chunk)
// plus the per-Server reasoning lookup threaded through the call chain
// instead of a process-global hook.
func (s *Server) convertOptions() convert.Options {
	opts := convert.Options{
		MaxSchemaNodes:          convert.DefaultMaxSchemaNodes,
		CompressKeepLast:        convert.DefaultCompressKeepLast,
		CompressMaxContentBytes: convert.DefaultCompressMaxContentBytes,
		ReasoningLookup: func(toolID, content, toolCallsJSON string) (string, string, bool) {
			if s.reasoningCache == nil {
				return "", "", false
			}
			return s.reasoningCache.Get(toolID, content, toolCallsJSON)
		},
	}
	if cfg := s.cfg.Load(); cfg != nil {
		opts.CompressPrompt = cfg.CompressPrompt
		opts.CacheControlInjection = cfg.CacheControlInjection
		opts.ReasoningInContent = cfg.ReasoningInContent
	}
	return opts
}

// WithVersion wires the running release tag + update checker for the
// dashboard badge (issue #50b). A nil checker disables the badge.
func WithVersion(version string, updates *updatecheck.Checker) Option {
	return func(s *Server) {
		s.version = version
		s.updates = updates
	}
}

// WithCommit wires the running build's commit hash (ARSYDONI UPDATE SOURCE,
// merge-guarded): the dashboard's primary update signal compares it against
// the repo's main-branch head, so any push to main marks the build outdated
// without a release. Empty (dev builds) degrades to the tag comparison.
func WithCommit(commit string) Option {
	return func(s *Server) {
		s.commit = commit
	}
}

// WithLoginClient wires the token-less upstream client that drives the
// headless OAuth login wizard (issue #62). A nil client disables the
// wizard endpoints with 503.
func WithLoginClient(c *upstream.Client) Option {
	return func(s *Server) {
		s.authClient = c
	}
}

// WithTokenDB wires the persistent SQLite-backed token store.  When set,
// add/remove mutations go through the database instead of rewriting .env.
func WithTokenDB(db *tokendb.DB) Option {
	return func(s *Server) {
		s.tokenDB = db
	}
}

// WithHistory threads the dashboard history store into the embedded admin
// UI (ADR-0016). A nil store keeps every history view on live data.
func WithHistory(st *store.Store) Option {
	return func(s *Server) {
		s.hist = st
	}
}

// Option configures optional server features (release-version badge).
type Option func(*Server)

// New builds the server over the configured pool and registry. A nil logger
// falls back to slog.Default(). The started timestamp pins /v1/models
// "created" and /healthz uptime. logs is the optional dashboard log viewer
// ring (nil disables the /admin/logs page data). configPath is the -config
// JSON path the process was started with ("" = none), used by reloads so a
// dashboard save or /admin/reload re-applies the JSON overrides. opts
// configure optional features (release-version badge, login wizard client).
func New(cfg *config.Config, p *pool.Pool, reg *registry.Registry, logger *slog.Logger, logs *logring.Handler, configPath string, opts ...Option) *Server {
	if logger == nil {
		logger = slog.Default()
	}
	s := &Server{pool: p, reg: reg, logger: logger, started: time.Now(), configPath: configPath, loginFlows: make(map[string]*loginFlow), logs: logs, gates: newAccessGates()}
	s.cfg.Store(cfg)
	s.rateLimiter = ratelimit.New(cfg.RateLimitPerIP, cfg.RateLimitBurst, 10000)
	// The token estimator shares one o200k_base codec process-wide, so
	// count_tokens requests never rebuild the vocabulary.
	est, err := tokenestimate.New()
	if err != nil {
		logger.Warn("token estimator unavailable; /v1/messages/count_tokens will fail", "err", err)
	}
	s.tokenEstimator = est
	for _, opt := range opts {
		opt(s)
	}
	if cfg.DashboardEnabled {
		dashOpts := []dashboard.Option{}
		if s.version != "" {
			dashOpts = append(dashOpts, dashboard.WithVersion(s.version, s.updates, s.commit))
		}
		if s.hist != nil {
			dashOpts = append(dashOpts, dashboard.WithHistory(s.hist))
		}
		s.dash = dashboard.New(func() *config.Config { return s.cfg.Load() }, p, reg, logger, logs, dashOpts...)
	}
	s.adminAuth = newAdminAuth()
	s.admin = &adminHandlers{
		dash:           s.dash,
		logfunc:        func() *slog.Logger { return s.logger },
		pool:           p,
		reg:            reg,
		cfgLoad:        s.cfg.Load,
		cfgStore:       s.cfg.Store,
		configPath:     configPath,
		settings:       s.hist,
		adminAuth:      s.adminAuth,
		loginFlows:     s.loginFlows,
		authClientFunc: func() *upstream.Client { return s.authClient },
		rateLimiter:    s.rateLimiter,
		handleChat:     s.handleChat,
		tokenDB:        s.tokenDB,
	}
	s.reasoningCache = reasoningcache.New(10000, 2*time.Hour)
	return s
}

// Close flushes and releases server-owned resources: the dashboard history
// consumer and store. Safe to call on a server built without WithHistory.
func (s *Server) Close() error {
	if s.dash != nil {
		return s.dash.Close()
	}
	return nil
}

// registerAdminRoutes mounts every dashboard.AdminRoutes row on the mux.
// Each row's Auth level selects the wrapping middleware stack it has always
// carried (see dashboard.AdminRoute for the level semantics); POST and
// DELETE rows are additionally wired through the CSRF gate. A row whose
// Path has no handler mapping panics — the table and the mapper ship as
// one commit.
func (s *Server) registerAdminRoutes(mux *http.ServeMux) {
	for _, r := range dashboard.AdminRoutes {
		h := s.adminHandler(r)
		// POST, PUT, and DELETE rows mutate state through the session cookie,
		// so all three carry the CSRF gate (GET rows are reads; the gate
		// itself skips GET/HEAD anyway).
		if r.Method == http.MethodPost || r.Method == http.MethodPut || r.Method == http.MethodDelete {
			h = s.admin.adminCSRF(h)
		}
		switch r.Auth {
		case dashboard.AuthNone:
			// No auth wrapper: login page, logout, static assets.
		case dashboard.AuthDashboard:
			h = s.admin.dashboardAuth(h)
		case dashboard.AuthSensitive:
			h = s.admin.adminSensitive(h)
			h = s.admin.dashboardAuth(h)
		case dashboard.AuthAdminToken:
			h = s.admin.adminSensitive(h)
			h = s.requireAdminToken(h)
		default:
			panic("server: unknown admin auth level " + r.Auth)
		}
		mux.Handle(r.Method+" "+r.Path, h)
	}
}

// adminHandler resolves one AdminRoutes row to its handler implementation,
// before auth wrapping. Unmapped paths panic: a table row without an
// implementation is a build invariant violation.
func (s *Server) adminHandler(r dashboard.AdminRoute) http.Handler {
	switch r.Method + " " + r.Path {
	case "POST /admin/reload":
		return http.HandlerFunc(s.admin.handleReload)
	case "GET /admin/login":
		return http.HandlerFunc(s.admin.handleAdminLogin)
	case "POST /admin/login":
		return http.HandlerFunc(s.admin.handleAdminLogin)
	case "POST /admin/logout":
		return http.HandlerFunc(s.admin.handleAdminLogout)
	case "GET /admin/api/overview":
		return s.dash.APIHandler("overview")
	case "GET /admin/api/tokens":
		return s.dash.APIHandler("tokens")
	case "GET /admin/api/events":
		return http.HandlerFunc(s.dash.HandleEvents)
	case "GET /admin/api/models":
		return s.dash.APIHandler("models")
	case "GET /admin/api/traces":
		return s.dash.APIHandler("traces")
	case "GET /admin/api/setup":
		return s.dash.APIHandler("setup")
	case "GET /admin/api/config":
		return s.dash.APIHandler("config")
	case "GET /admin/api/config/meta":
		return http.HandlerFunc(s.dash.APIConfigMeta)
	case "GET /admin/api/settings":
		return http.HandlerFunc(s.admin.handleSettingsGet)
	case "POST /admin/api/settings":
		return http.HandlerFunc(s.admin.handleSettingsPost)
	case "DELETE /admin/api/settings/{key}":
		return http.HandlerFunc(s.admin.handleSettingsDelete)
	case "GET /admin/api/pages/{id}":
		return http.HandlerFunc(s.admin.handlePageStateGet)
	case "PUT /admin/api/pages/{id}":
		return http.HandlerFunc(s.admin.handlePageStatePut)
	case "GET /admin/api/logs":
		return s.dash.APIHandler("logs")
	case "GET /admin/api/quota/history":
		return s.dash.APIHandler("quota/history")
	case "GET /admin/api/maturity/history":
		return s.dash.APIHandler("maturity/history")
	case "GET /admin/api/logs/history":
		return s.dash.APIHandler("logs/history")
	case "GET /admin/api/logs/rollup":
		return http.HandlerFunc(s.dash.APILogsRollup)
	case "GET /admin/api/logs/export":
		return http.HandlerFunc(s.dash.APILogsExport)
	case "POST /admin/api/logs/import":
		return http.HandlerFunc(s.dash.APILogsImport)
	case "GET /admin/api/metrics":
		return s.dash.APIHandler("metrics")
	case "GET /admin/api/usage":
		return s.dash.APIHandler("usage")
	case "GET /admin/api/version":
		return http.HandlerFunc(s.dash.APIVersion)
	case "GET /admin/api/auth/status":
		return http.HandlerFunc(s.admin.handleAdminAuthStatus)
	case "GET /admin/api/notices":
		return s.dash.APIHandler("notices")
	case "GET /admin", "GET /admin/", "GET /admin/tokens", "GET /admin/models", "GET /admin/traces",
		"GET /admin/setup", "GET /admin/config", "GET /admin/logs", "GET /admin/metrics":
		// SPA shell routes: the gateway serves the Svelte app directly.
		return http.HandlerFunc(s.dash.ServeSPA)
	case "GET /admin/playground":
		// Server-side DEVTOOLS_ENABLED gate: the playground shell is a
		// Dev Tools page; a direct GET must 404 when the knob is off.
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			if !s.cfg.Load().DevToolsEnabled {
				s.dash.RenderResult(w, http.StatusNotFound, false, "dev tools are disabled — set DEVTOOLS_ENABLED=true to enable the playground", "devtools_disabled")
				return
			}
			s.dash.ServeSPA(w, r)
		})
	case "POST /admin/playground/chat":
		return http.HandlerFunc(s.admin.handlePlaygroundChat)
	case "POST /admin/login/start":
		return http.HandlerFunc(s.admin.handleLoginStart)
	case "GET /admin/login/status":
		return http.HandlerFunc(s.admin.handleLoginStatus)
	case "POST /admin/config":
		return http.HandlerFunc(s.admin.handleConfigSave)
	case "POST /admin/tokens/{id}/unlock":
		return http.HandlerFunc(s.admin.handleTokenUnlock)
	case "POST /admin/tokens/{id}/lock":
		return http.HandlerFunc(s.admin.handleTokenLock)
	case "POST /admin/tokens/{id}/unlock-lock":
		return http.HandlerFunc(s.admin.handleTokenUnlockLock)
	case "POST /admin/tokens/{id}/maturity":
		return http.HandlerFunc(s.admin.handleTokenMaturity)
	case "POST /admin/tokens/{id}/maturity/touch":
		return http.HandlerFunc(s.admin.handleTokenMaturityTouch)
	case "POST /admin/bridge-tokens/{key}/lock":
		return http.HandlerFunc(s.admin.handleBridgeTokenLock)
	case "POST /admin/bridge-tokens/{key}/unlock":
		return http.HandlerFunc(s.admin.handleBridgeTokenUnlock)
	case "POST /admin/tokens/{id}/finish":
		return http.HandlerFunc(s.admin.handleTokenFinish)
	case "POST /admin/tokens/{id}/drop-session":
		return http.HandlerFunc(s.admin.handleTokenDropSession)
	case "POST /admin/tokens/{id}/refund-refresh":
		return http.HandlerFunc(s.admin.handleTokenRefundRefresh)
	case "POST /admin/tokens/{id}/test":
		return http.HandlerFunc(s.admin.handleTokenTest)
	case "POST /admin/tokens/test-all":
		return http.HandlerFunc(s.admin.handleTokensTestAll)
	case "POST /admin/tokens/{id}/session":
		return http.HandlerFunc(s.admin.handleTokenSpawnSession)
	case "POST /admin/tokens/add":
		return http.HandlerFunc(s.admin.handleTokenAdd)
	case "POST /admin/tokens/remove":
		return http.HandlerFunc(s.admin.handleTokenRemove)
	case "POST /admin/tokens/swap":
		return http.HandlerFunc(s.admin.handleTokenSwap)
	case "POST /admin/mode":
		return http.HandlerFunc(s.admin.handleModeSwitch)
	case "POST /admin/diag":
		return http.HandlerFunc(s.admin.handleDiag)
	case "POST /admin/api/change-password":
		return http.HandlerFunc(s.admin.handleAdminChangePassword)
	case "POST /admin/api/require-login":
		return http.HandlerFunc(s.admin.handleAdminRequireLogin)
	case "POST /admin/smoke":
		return http.HandlerFunc(s.admin.handleSmoke)
	case "POST /admin/restart":
		return http.HandlerFunc(s.admin.handleAdminRestart)
	case "GET /admin/assets/":
		return noDirListing(http.StripPrefix("/admin/assets/", http.FileServerFS(mustSubFS(dashboard.DistFS(), "assets"))))
	case "POST /admin/tokens/remove-specific":
		return http.HandlerFunc(s.handleTokenRemoveSpecific)
	case "GET /admin/tokens/list":
		return http.HandlerFunc(s.handleTokenList)
	default:
		panic("server: no admin handler for " + r.Method + " " + r.Path)
	}
}

// handleRoot answers GET /: 302 to /admin when the dashboard is enabled, 404
// otherwise. The exact-root pattern leaves every other path on the ServeMux
// default (/v1/*, /healthz, /metrics, /admin/* contracts untouched).
func (s *Server) handleRoot(w http.ResponseWriter, r *http.Request) {
	if !s.cfg.Load().DashboardEnabled {
		http.NotFound(w, r)
		return
	}
	http.Redirect(w, r, "/admin", http.StatusFound)
}

// Handler returns the route table wrapped in an access-log middleware. Method
// mismatches and unknown paths get the ServeMux's automatic 405/404.
func (s *Server) Handler() http.Handler {
	mux := http.NewServeMux()
	s.registerOpenAIRoutes(mux)
	s.registerAnthropicRoutes(mux)
	mux.HandleFunc("GET /healthz", s.handleHealthz)
	mux.HandleFunc("GET /metrics", s.handleMetrics)
	mux.HandleFunc("GET /{$}", s.handleRoot)
	if s.cfg.Load().DashboardEnabled {
		s.registerAdminRoutes(mux)
	}
	// CORS middleware wraps the whole route table: it answers OPTIONS
	// preflights on the /v1/* API surface with 204 and stamps the allow
	// headers on every /v1/* response. Admin routes are intentionally left
	// untouched (cookie-authenticated dashboard; SameSite=Strict already
	// blocks cross-site reads, and an allow-origin would add nothing there).
	cors := s.corsMiddleware(mux)
	h := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		start := time.Now()
		// D1: mint the request's correlation id exactly once here, then
		// carry it in the request context so every downstream log line
		// (chat routing/done/trace, request failed, upstream do/retry)
		// shares it. Handlers reached without this wrapper (direct calls
		// in tests) mint a fallback id in chatCore.
		reqID := newReqID()
		ctx, accessTok := withAccessToken(r.Context())
		r = r.WithContext(context.WithValue(ctx, reqIDKey{}, reqID))
		// Echo the minted id so callers can join client-side traces with
		// gateway logs. Set before any downstream write (headers flush on
		// the first WriteHeader) so EVERY path through this wrapper
		// carries it — logged and silent (healthz/probes, dashboard GET
		// polls) alike, plus the early rate-limit refusal below. The
		// inbound X-Request-Id is never adopted (spoofable); it is only
		// logged as client_request_id.
		w.Header().Set("X-Request-Id", reqID)
		// Client-side per-IP rate limiting at the OUTERMOST wrapper (issue
		// #137): when RATE_LIMIT_PER_IP is enabled every /v1/* route is
		// covered — chat completions, Responses, Anthropic messages,
		// count_tokens, /v1/models — not just the completion core, so a
		// client cannot burn upstream work through an unthrottled surface.
		// Exempt (documented): /admin/* (the dashboard owns its own
		// throttles), /healthz and /metrics (liveness/monitoring must never
		// be rate-limited away), CORS OPTIONS preflights (they must answer
		// so browsers learn the policy), and non-/v1/ paths (404s are
		// logged only, they cost no upstream work).
		cfg := s.cfg.Load()
		if cfg.RateLimitPerIP > 0 && strings.HasPrefix(r.URL.Path, "/v1/") && r.Method != http.MethodOptions {
			if allowed, retryAfter := s.rateLimiter.Allow(r.RemoteAddr); !allowed {
				retrySec := int(math.Ceil(retryAfter.Seconds()))
				if retrySec < 1 {
					retrySec = 1
				}
				s.logger.Warn("rate limit exceeded",
					"remote", remoteHost(r),
					"req_id", reqID,
					"retry_after_sec", retrySec,
				)
				s.rateLimitRejections.Add(1)
				// One envelope dispatch (issue #253): the wire decides the
				// body shape; the Retry-After ceiling is computed inside.
				s.writeClientError(w, r, http.StatusTooManyRequests,
					fmt.Sprintf("client rate limit exceeded (Retry-After: %ds)", retrySec),
					"rate_limit_exceeded", retryAfter)
				return
			}
		}
		sw := &statusWriter{ResponseWriter: w, status: http.StatusOK}
		cors.ServeHTTP(sw, r)
		attrs := []any{
			"req_id", reqID,
			"method", r.Method,
			"path", r.URL.Path,
			"status", sw.status,
			"ms", time.Since(start).Milliseconds(),
			"remote", remoteHost(r),
		}
		// The client's X-Request-Id is preserved as a separate
		// client_request_id field (never trusted as the correlation key).
		if crid := clientRequestID(r); crid != "" {
			attrs = append(attrs, "client_request_id", crid)
		}
		// Token identity for abusive-key triage from access lines alone:
		// chatCore stashed the serving lease's label (1-based index or
		// "bridge", never the raw key) into the carrier during the
		// request; non-chat surfaces and pre-lease refusals leave it
		// empty and the field stays absent (ring-only there).
		if accessTok.token != "" {
			attrs = append(attrs, "token", accessTok.token)
		}
		// T17: LOG_ACCESS=false disables access lines entirely. Silent
		// paths (probes + dashboard GET polls) never log; quiet endpoints
		// (OPTIONS preflights) and UNKNOWN paths (404s) are rate-limited
		// to one access line per path per accessQuietWindow, and the
		// quiet-class budget caps the total quiet lines per window — a
		// client minting distinct paths cannot grow the access log without
		// bound (per-path gating alone would pass one line per unique
		// path). req_id/client_request_id survive in both cases.
		if !cfg.LogAccess {
			return
		}
		if silentAccessPath(r.Method, r.URL.Path) {
			return
		}
		if quiet := quietAccessPath(r.Method, r.URL.Path) || sw.status == http.StatusNotFound; quiet {
			if !s.gates.accessLogDue(r.URL.Path, start) || !s.gates.accessQuietBudgetDue(start) {
				return
			}
		}
		s.logger.Info("access", attrs...)
	})
	// gzip wraps the ENTIRE route table (admin + APIs). Streaming
	// responses are exempted inside by Content-Type gate; HEAD and
	// no-body statuses are skipped.
	return gzipMiddleware(h)
}